fix currency formatting to preserve decimal values
Replace parseInt with parseFloat in formatCurrency across default, fractional, land-banking, and outright property pages to avoid truncating decimal amounts.

+++ src/routes/investors/properties/$propertyId/fractional/index.tsx
@@ -37,7 +37,7 @@ function PropertyDetailPage() {
   });
   const formatCurrency = (amount: number | null | undefined) => {
     if (amount === null || amount === undefined) return "N/A";
-    const fixed = parseInt(amount.toPrecision());
+    const fixed = parseFloat(amount.toPrecision());
     return `₦ ${fixed.toLocaleString()}`;
   };
   const mutate = useMutation({
@@ -595,7 +595,7 @@ const InstallMentForm = ({
 }) => {
   const formatCurrency = (amount: number | null | undefined) => {
     if (amount === null || amount === undefined) return "N/A";
-    const fixed = parseInt(amount.toPrecision());
+    const fixed = parseFloat(amount.toPrecision());
     return `₦ ${fixed.toLocaleString()}`;
   };
