+++ src/routes/investors/properties/$propertyId/outright/index.tsx
@@ -44,7 +44,7 @@ function PropertyDetailPage() {
   });
   const formatCurrency = (amount: number | null | undefined) => {
     if (amount === null || amount === undefined) return "N/A";
-    const fixed = parseInt(amount.toPrecision());
+    const fixed = parseFloat(amount.toPrecision());
     return `₦ ${fixed.toLocaleString()}`;
   };
   const mutate = useMutation({
@@ -517,7 +517,7 @@ const InstallMentForm = ({
 }) => {
   const formatCurrency = (amount: number | null | undefined) => {
     if (amount === null || amount === undefined) return "N/A";
-    const fixed = parseInt(amount.toPrecision());
+    const fixed = parseFloat(amount.toPrecision());
     return `₦ ${fixed.toLocaleString()}`;
   };
